Finalize 32-step upgrade: UI refinements, pattern fixes, and dynamic default sequences

+++ eeprom.h
@@ -4,9 +4,11 @@
 
 void eeprom_init();
 
-void eeprom_write_pattern(uint8_t slot, const uint8_t* notes, uint16_t gate_mask, uint8_t steps);
+void eeprom_write_pattern(uint8_t slot, const uint8_t *notes,
+                          uint32_t gate_mask, uint8_t steps);
 
-void eeprom_read_pattern(uint8_t slot, uint8_t* notes, uint16_t* gate_mask, uint8_t* steps);
+void eeprom_read_pattern(uint8_t slot, uint8_t *notes, uint32_t *gate_mask,
+                         uint8_t *steps);
 
 bool eeprom_is_initialized();
